api/requirements.txt
azure-functions==1.21.3
google-genai==1.14.0
